moved samples to helpers
