+++ api/app/spectrogram.py
@@ -2,7 +2,7 @@ import io
 import matplotlib.pyplot as plt
 import numpy as np
 from PIL import Image
-from .samples import get_samples
+from helpers.samples import get_samples
 
 def get_spectrogram_image(
     content: bytes,
