+++ api/tests/test_samples.py
@@ -1,6 +1,6 @@
 import numpy as np
 import pytest
-from app.samples import get_samples
+from helpers.samples import get_samples
 
 
 def test_get_samples_i8():
